+++ particula/activity/gibbs.py
@@ -1,4 +1,5 @@
-"""Calculates the Gibbs free energy of mixing for a binary solution."""
+"""Calculates the Gibbs free energy of mixing for a binary solution.
+"""
 
 from particula.util.machine_limit import safe_log
