+++ particula/activity/phase_separation.py
@@ -1,22 +1,17 @@
 """
 This module contains functions to calculate the phase separation of organic
-compounds in water. The functions are based on the BAT model by Gorkowski
-et al. (2019).
+compounds in water. The functions are based on the BAT model.
 
-Gorkowski, K., Preston, T. C., &#38; Zuend, A. (2019).
+Gorkowski, K., Preston, T. C., Zuend, A. (2019).
 Relative-humidity-dependent organic aerosol thermodynamics
 Via an efficient reduced-complexity model.
 Atmospheric Chemistry and Physics
 https://doi.org/10.5194/acp-19-13383-2019
 """
 
-from typing import Union, Optional, List
+from typing import Union
 import numpy as np
 from numpy.typing import NDArray
-from particula.activity.bat_coefficents import INTERPOLATE_WATER_FIT, LOWEST_ORGANIC_MOLE_FRACTION
-from particula.activity.activity_coefficients import activity_coefficients
-from particula.activity.species_density import organic_density_estimate
-from particula.activity import phase_separation
 
 from particula.util.machine_limit import safe_exp, safe_log
 
@@ -49,15 +44,14 @@ def organic_water_single_phase(
     molar_mass_ratio = np.asarray(molar_mass_ratio, dtype=np.float64)
 
     return (
-        0.205
-        / (1 + safe_exp(26.6 * (molar_mass_ratio - 0.12))) ** 0.843
+        0.205 / (1 + safe_exp(26.6 * (molar_mass_ratio - 0.12))) ** 0.843
         + 0.225
     )
 
 
 # pylint: disable=too-many-locals
 def find_phase_sep_index(
-        activity_data: ArrayLike
+    activity_data: NDArray[np.float64]
 ) -> dict:
     """
     This function finds phase separation using activity>1 and
@@ -94,8 +88,7 @@ def find_phase_sep_index(
             # If so, no phase separation via activity curvature
             phase_sep_curve = 0
             # find the index where the activity is closest to 1
-            index_phase_sep_starts = np.argmin(
-                np.abs(activity_data - 1))
+            index_phase_sep_starts = np.argmin(np.abs(activity_data - 1))
             index_phase_sep_end = index_phase_sep_starts
         else:
             # If signs differ, phase separation via activity curvature occurs
@@ -105,23 +98,32 @@ def find_phase_sep_index(
 
             # The first index where a sign change occurs
             inflection_index = np.where(sign_changes)[0]  # all indices
-            index_start = inflection_index[0] \
-                if len(inflection_index) > 0 else data_length
+            index_start = (
+                inflection_index[0]
+                if len(inflection_index) > 0
+                else data_length
+            )
             # The last index where a sign change occurs
-            back_index = inflection_index[-1] \
-                if len(inflection_index) > 0 else data_length
+            back_index = (
+                inflection_index[-1]
+                if len(inflection_index) > 0
+                else data_length
+            )
 
             # Check if first section of activity data is greater than 1
             if np.any(activity_data[:index_start] > 1):
                 index_phase_sep_starts = np.argmin(
-                    np.abs(activity_data[:index_start] - 1))
+                    np.abs(activity_data[:index_start] - 1)
+                )
             else:
                 index_phase_sep_starts = index_start
 
             # Check if second section of activity data is greater than 1
             if np.any(activity_data[back_index:] > 1):
-                index_phase_sep_end = np.argmin(
-                    np.abs(activity_data[back_index:] - 1)) + back_index
+                index_phase_sep_end = (
+                    np.argmin(np.abs(activity_data[back_index:] - 1))
+                    + back_index
+                )
             else:
                 index_phase_sep_end = back_index
     else:
@@ -132,15 +134,17 @@ def find_phase_sep_index(
 
     # Assign phase separation via activity based on data being greater than 1
     phase_sep_activity = 1 if sum(activity_data > 1) else 0
-    return {'phase_sep_activity': phase_sep_activity,
-            'phase_sep_curve': phase_sep_curve,
-            'index_phase_sep_starts': index_phase_sep_starts,
-            'index_phase_sep_end': index_phase_sep_end}
+    return {
+        "phase_sep_activity": phase_sep_activity,
+        "phase_sep_curve": phase_sep_curve,
+        "index_phase_sep_starts": index_phase_sep_starts,
+        "index_phase_sep_end": index_phase_sep_end,
+    }
 
 
 def find_phase_separation(
-        activity_water: ArrayLike,
-        activity_org: ArrayLike
+    activity_water: NDArray[np.float64],
+    activity_org: NDArray[np.float64]
 ) -> dict:
     """
     This function checks for phase separation in each activity curve.
@@ -166,22 +170,21 @@ def find_phase_separation(
         - 'matching_upper_seperation': The value of water activity at the
                 matching upper separation point.
     """
-    # check inputs
-    activity_water = np.asarray(activity_water, dtype=np.float64)
-    activity_org = np.asarray(activity_org, dtype=np.float64)
 
     # check for phase separation in each activity curve
     water_sep = find_phase_sep_index(activity_water)
     organic_sep = find_phase_sep_index(activity_org)
 
     # gather all the indexes into a list for easier access
-    indexes = [water_sep['index_phase_sep_starts'],
-               water_sep['index_phase_sep_end'],
-               organic_sep['index_phase_sep_starts'],
-               organic_sep['index_phase_sep_end']]
+    indexes = [
+        water_sep["index_phase_sep_starts"],
+        water_sep["index_phase_sep_end"],
+        organic_sep["index_phase_sep_starts"],
+        organic_sep["index_phase_sep_end"],
+    ]
 
     # If there is a phase separation curve in the water activity data
-    if water_sep['phase_sep_curve'] == 1:
+    if water_sep["phase_sep_curve"] == 1:
         phase_sep_check = 1
 
         # Check for the direction of the curve (increasing or decreasing)
@@ -200,8 +203,9 @@ def find_phase_separation(
             if len(match_index_prime[0]) == 0:
                 match_index_prime = lower_seperation_index
             else:
-                match_index_prime = match_index_prime[0][0] + \
-                    lower_seperation_index
+                match_index_prime = (
+                    match_index_prime[0][0] + lower_seperation_index
+                )
         else:  # decreasing a_w with index
             # find the min and max indexes
             lower_seperation_index = max(indexes)
@@ -225,18 +229,20 @@ def find_phase_separation(
         match_index_prime = 2
         phase_sep_check = 0  # no phase seperation
 
-    return {'phase_sep_check': phase_sep_check,
-            'lower_seperation_index': lower_seperation_index,
-            'upper_seperation_index': upper_seperation_index,
-            'matching_upper_seperation_index': match_index_prime,
-            'lower_seperation': activity_water[lower_seperation_index],
-            'upper_seperation': activity_water[upper_seperation_index],
-            'matching_upper_seperation': activity_water[match_index_prime]}
+    return {
+        "phase_sep_check": phase_sep_check,
+        "lower_seperation_index": lower_seperation_index,
+        "upper_seperation_index": upper_seperation_index,
+        "matching_upper_seperation_index": match_index_prime,
+        "lower_seperation": activity_water[lower_seperation_index],
+        "upper_seperation": activity_water[upper_seperation_index],
+        "matching_upper_seperation": activity_water[match_index_prime],
+    }
 
 
 def q_alpha(
-        seperation_activity: Union[int, float, np.ndarray],
-        activities: ArrayLike,
+    seperation_activity: NDArray[np.float64],
+    activities: NDArray[np.float64],
 ) -> np.ndarray:
     """
     This function calculates the q_alpha value using a squeezed logistic
@@ -257,161 +263,29 @@ def q_alpha(
             q_alpha value.
     """
     # check inputs
-    activities = np.asarray(activities, dtype=np.float64)
     if seperation_activity == 0:
         return np.ones_like(activities)
-    seperation_activity = np.asarray(seperation_activity, dtype=np.float64)
 
     # spread in transfer from 50/50 point
     delta_seperation = 1 - seperation_activity
 
     # check min value allowed
     above_min_delta_seperation_value = delta_seperation > MIN_SPREAD_IN_AW
-    delta_seperation = delta_seperation * above_min_delta_seperation_value + \
-        ~above_min_delta_seperation_value * MIN_SPREAD_IN_AW
+    delta_seperation = (
+        delta_seperation * above_min_delta_seperation_value
+        + ~above_min_delta_seperation_value * MIN_SPREAD_IN_AW
+    )
 
     # calculate curve parameter of sigmoid
-    sigmoid_curve_parameter = safe_log(
-        1 / (1 - Q_ALPHA_AT_1PHASE_AW) - 1) / delta_seperation
+    sigmoid_curve_parameter = (
+        safe_log(1 / (1 - Q_ALPHA_AT_1PHASE_AW) - 1) / delta_seperation
+    )
 
     # calculate q_alpha return value
     return 1 - 1 / (
         1
         + safe_exp(
-            sigmoid_curve_parameter *
-            (activities - seperation_activity + delta_seperation)
+            sigmoid_curve_parameter
+            * (activities - seperation_activity + delta_seperation)
         )
     )
-
-
-def biphasic_water_activity_point(
-    oxygen2carbon: Union[float, NDArray[np.float64]],
-    hydrogen2carbon: Union[float, NDArray[np.float64]],
-    molar_mass_ratio: Union[float, NDArray[np.float64]],
-    functional_group: Optional[Union[list[str], str]] = None,
-) -> np.ndarray:
-    """
-    This function computes the biphasic to single phase
-    water activity (RH*100).
-
-    Args:
-        - oxygen2carbon : The oxygen to carbon ratio.
-        - hydrogen2carbon : The hydrogen to carbon ratio.
-        - molar_mass_ratio : The molar mass ratio of water to organic
-          matter.
-        - functional_group : Optional functional group(s) of the organic
-          compound, if applicable.
-
-    Returns:
-        - The RH cross point array.
-    """
-    oxygen2carbon = np.asarray(oxygen2carbon, dtype=np.float64)
-    hydrogen2carbon = np.asarray(hydrogen2carbon, dtype=np.float64)
-    molar_mass_ratio = np.asarray(molar_mass_ratio, dtype=np.float64)
-    if oxygen2carbon.ndim == 0:
-        oxygen2carbon = np.expand_dims(oxygen2carbon, axis=0)
-    if hydrogen2carbon.ndim == 0:
-        hydrogen2carbon = np.expand_dims(hydrogen2carbon, axis=0)
-    if molar_mass_ratio.ndim == 0:
-        molar_mass_ratio = np.expand_dims(molar_mass_ratio, axis=0)
-
-    water_activity_cross_point = np.zeros_like(oxygen2carbon)
-
-    interpolate_step_numb = 200
-    mole_frac = np.logspace(-6, 0, interpolate_step_numb + 1)
-
-    for i, _ in enumerate(oxygen2carbon):
-        density = organic_density_estimate(
-            molar_mass_ratio[i],
-            oxygen2carbon[i],
-            hydrogen2carbon[i],
-            mass_ratio_convert=True,
-        )
-        activities = activity_coefficients(
-            molar_mass_ratio=molar_mass_ratio[i],
-            organic_mole_fraction=mole_frac,
-            oxygen2carbon=oxygen2carbon[i],
-            density=density,
-            functional_group=functional_group,
-        )
-
-        if np.isnan(activities[0]).any():
-            raise ValueError("water activity is NaN, check inputs")
-
-        phase_check = phase_separation.find_phase_separation(
-            activities[0], activities[1]
-        )
-
-        if phase_check["phase_sep_check"] == 1:
-            water_activity_cross_point[i] = phase_check["upper_seperation"]
-        else:
-            water_activity_cross_point[i] = 0
-
-    water_activity_cross_point[water_activity_cross_point < 0] = 0
-    water_activity_cross_point[water_activity_cross_point > 1] = 1
-
-    return water_activity_cross_point
-
-
-def fixed_water_activity(
-    water_activity: Union[float, NDArray[np.float64]],
-    molar_mass_ratio: Union[float, NDArray[np.float64]],
-    oxygen2carbon: Union[float, NDArray[np.float64]],
-    density: Union[float, NDArray[np.float64]],
-) -> Tuple[
-    Union[float, NDArray[np.float64]],
-    Union[float, NDArray[np.float64]],
-    Union[float, NDArray[np.float64]],
-]:
-    """
-    Calculate the activity coefficients of water and organic matter in
-    organic-water mixtures.
-
-    This function assumes a fixed water activity value (e.g., RH = 75%
-    corresponds to 0.75 water activity in equilibrium).
-    It calculates the activity coefficients for different phases and
-    determines phase separations if they occur.
-
-    Args:
-        - water_activity : An array of water activity values.
-        - molar_mass_ratio : Array of molar mass ratios of the components.
-        - oxygen2carbon : Array of oxygen-to-carbon ratios.
-        - density : Array of densities of the mixture, in kg/m^3.
-
-    Returns:
-        - A tuple containing the activity coefficients for alpha and beta
-          phases, and the q_alpha (phase separation) value.
-          If no phase separation occurs, the beta phase values are None.
-    """
-    water_activity = np.asarray(water_activity, dtype=np.float64)
-    molar_mass_ratio = np.asarray(molar_mass_ratio, dtype=np.float64)
-    oxygen2carbon = np.asarray(oxygen2carbon, dtype=np.float64)
-    density = np.asarray(density, dtype=np.float64)
-
-    if water_activity.size > 1 and water_activity[0] > water_activity[-1]:
-        water_activity = np.flip(water_activity)
-        flip = True
-    else:
-        flip = False
-
-    organic_mole_fraction_array = np.linspace(
-        1,
-        LOWEST_ORGANIC_MOLE_FRACTION,
-        INTERPOLATE_WATER_FIT,
-        dtype=np.float64,
-    )
-
-    activities = activity_coefficients(
-        molar_mass_ratio=molar_mass_ratio,
-        organic_mole_fraction=organic_mole_fraction_array,
-        oxygen2carbon=oxygen2carbon,
-        density=density,
-    )
-    phase_check = phase_separation.find_phase_separation(
-        activities[0], activities[1]
-    )
-    activities_water = np.asarray(activities[0], dtype=np.float64)
-    if phase_check["phase_sep_check"] == 0:
-        alpha_organic_mole_fraction = np.interp(
-            xp=activities_water,
-            fp=organic_mole_fraction_array,
